Move edit route tests into edit_spec, remove edit_inertia_spec
